+++ jabsrv/src/main/java/module-info.java
@@ -58,5 +58,4 @@ module org.jabref.jabsrv {
     requires jersey.server;
     requires com.fasterxml.jackson.annotation;
     requires com.fasterxml.jackson.databind;
-
 }
