+++ jabsrv/src/main/java/org/jabref/http/server/LibraryResource.java
@@ -114,115 +114,9 @@ public class LibraryResource {
                        .build();
     }
 
-    /**
-     * @param id - also "demo" for the Chocolate.bib file
-     */
-    private BibDatabaseContext getDatabaseContext(String id) throws IOException {
-        return ServerUtils.getBibDatabaseContext(id, filesToServe, srvStateManager, preferences.getImportFormatPreferences());
-    }
-
-    /**
-     * At http://localhost:23119/libraries/{id}/entries/{entryId} <br><br>
-     * <p>
-     * Combines attributes of a given BibEntry into a basic entry preview for as plain text.
-     *
-     * @param id      The name of the library
-     * @param entryId The CitationKey of the BibEntry
-     * @return a basic entry preview as plain text
-     * @throws IOException
-     * @throws NotFoundException
-     */
-    @GET
-    @Path("entries/{entryId}")
-    @Produces(MediaType.TEXT_PLAIN + ";charset=UTF-8")
-    public String getPlainRepresentation(@PathParam("id") String id, @PathParam("entryId") String entryId) throws IOException {
-        BibDatabaseContext databaseContext = getDatabaseContext(id);
-        List<BibEntry> entriesByCitationKey = databaseContext.getDatabase().getEntriesByCitationKey(entryId);
-        if (entriesByCitationKey.isEmpty()) {
-            throw new NotFoundException("Entry with citation key '" + entryId + "' not found in library " + id);
-        }
-        if (entriesByCitationKey.size() > 1) {
-            LOGGER.warn("Multiple entries found with citation key '{}'. Using the first one.", entryId);
-        }
-
-        // TODO: Currently, the preview preferences are in GUI package, which is not accessible here.
-        // build the preview
-        BibEntry entry = entriesByCitationKey.getFirst();
-
-        String author = entry.getField(StandardField.AUTHOR).orElse("(N/A)");
-        String title = entry.getField(StandardField.TITLE).orElse("(N/A)");
-        String journal = entry.getField(StandardField.JOURNAL).orElse("(N/A)");
-        String volume = entry.getField(StandardField.VOLUME).orElse("(N/A)");
-        String number = entry.getField(StandardField.NUMBER).orElse("(N/A)");
-        String pages = entry.getField(StandardField.PAGES).orElse("(N/A)");
-        String releaseDate = entry.getField(StandardField.DATE).orElse("(N/A)");
-
-        // the only difference to the HTML version of this method is the format of the output:
-        String preview =
-                "Author: " + author
-                        + "\nTitle: " + title
-                        + "\nJournal: " + journal
-                        + "\nVolume: " + volume
-                        + "\nNumber: " + number
-                        + "\nPages: " + pages
-                        + "\nReleased on: " + releaseDate;
-
-        return preview;
-    }
-
-    /**
-     * At http://localhost:23119/libraries/{id}/entries/{entryId} <br><br>
-     * <p>
-     * Combines attributes of a given BibEntry into a basic entry preview for as HTML text.
-     *
-     * @param id      The name of the library
-     * @param entryId The CitationKey of the BibEntry
-     * @return a basic entry preview as HTML text
-     * @throws IOException
-     */
-    @GET
-    @Path("entries/{entryId}")
-    @Produces(MediaType.TEXT_HTML + ";charset=UTF-8")
-    public String getHTMLRepresentation(@PathParam("id") String id, @PathParam("entryId") String entryId) throws IOException {
-        List<BibEntry> entriesByCitationKey = getDatabaseContext(id).getDatabase().getEntriesByCitationKey(entryId);
-        if (entriesByCitationKey.isEmpty()) {
-            throw new NotFoundException("Entry with citation key '" + entryId + "' not found in library " + id);
-        }
-        if (entriesByCitationKey.size() > 1) {
-            LOGGER.warn("Multiple entries found with citation key '{}'. Using the first one.", entryId);
-        }
-
-        // TODO: Currently, the preview preferences are in GUI package, which is not accessible here.
-        // build the preview
-        BibEntry entry = entriesByCitationKey.getFirst();
-
-        String author = entry.getField(StandardField.AUTHOR).orElse("(N/A)");
-        String title = entry.getField(StandardField.TITLE).orElse("(N/A)");
-        String journal = entry.getField(StandardField.JOURNAL).orElse("(N/A)");
-        String volume = entry.getField(StandardField.VOLUME).orElse("(N/A)");
-        String number = entry.getField(StandardField.NUMBER).orElse("(N/A)");
-        String pages = entry.getField(StandardField.PAGES).orElse("(N/A)");
-        String releaseDate = entry.getField(StandardField.DATE).orElse("(N/A)");
-
-        // the only difference to the plain text version of this method is the format of the output:
-        String preview =
-                "<strong>Author:</strong> " + author + "<br>" +
-                        "<strong>Title:</strong> " + title + "<br>" +
-                        "<strong>Journal:</strong> " + journal + "<br>" +
-                        "<strong>Volume:</strong> " + volume + "<br>" +
-                        "<strong>Number:</strong> " + number + "<br>" +
-                        "<strong>Pages:</strong> " + pages + "<br>" +
-                        "<strong>Released on:</strong> " + releaseDate;
-
-        return preview;
-    }
-
-    /**
-     * At http://localhost:23119/libraries/{id}/entries/pdffiles <br><br>
-     * <p>
-     * Loops through all entries in the specified library and adds attached files of type "PDF" to
-     * a list and JSON serialises it.
-     */
+    /// Loops through all entries in the specified library and adds attached files of type "PDF" to
+    /// a list and JSON serialises it.
+    /// FIXME: JabMap should serve the files per BibEntry. See <https://github.com/JabRef/jabmap/issues/56> for details
     @GET
     @Path("entries/pdffiles")
     @Produces(MediaType.APPLICATION_JSON + ";charset=UTF-8")
@@ -258,5 +152,10 @@ public class LibraryResource {
      */
     private @Nullable InputStream getChocolateBibAsStream() {
         return BibDatabase.class.getResourceAsStream("/Chocolate.bib");
+    }
+
+    /// @param id - also "demo" for the Chocolate.bib file
+    private BibDatabaseContext getDatabaseContext(String id) throws IOException {
+        return ServerUtils.getBibDatabaseContext(id, filesToServe, srvStateManager, preferences.getImportFormatPreferences());
     }
 }
